Change iteritems() to items()
Not sure if this project aims for Python 2 or 3 but I'm using 3 and kept getting an error “ 'dict' object has no attribute 'iteritems' ”.  Looked around and found this helpful page http://stackoverflow.com/questions/30418481/error-dict-object-has-no-attribute-iteritems-when-trying-to-use-networkx

Tried the change and everything works great now.  Looking at the python documentation, it -looks- like the change is backwards compatible for Python 2 users but I'm still new so don't quote me on that.